# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface KeeperStatus {
  configured: boolean;
  running: boolean;
  /** @nullable */
  operatorAddress: string | null;
  /** @nullable */
  contractAddress: string | null;
  /** @minimum 0 */
  activeMandates: number;
  /** @nullable */
  lastPollAt: string | null;
  /** @nullable */
  lastTransactionHash: string | null;
  /** @nullable */
  lastError: string | null;
}

export interface MarketPrice {
  symbol: string;
  /** @minimum 0 */
  priceUsd: number;
  /** @nullable */
  change24h: number | null;
}

export interface MarketPricesPayload {
  prices: MarketPrice[];
  updatedAt: string;
}

export interface MintSimulationInput {
  /** @pattern ^0x[a-fA-F0-9]{40}$ */
  from: string;
  /**
     * @maxLength 4096
     * @pattern ^0x[a-fA-F0-9]+$
     */
  data: string;
}

export interface MintSimulationResult {
  /** @pattern ^0x[a-fA-F0-9]*$ */
  result: string;
}