# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chain/simulate-mint:
    post:
      operationId: simulateMint
      tags: [chain]
      summary: Simulate a position mint
      description: Simulates a mint against the configured position manager without submitting a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MintSimulationInput"
      responses:
        "200":
          description: Encoded mint simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintSimulationResult"
        "400":
          description: Invalid simulation request
        "502":
          description: Chain RPC rejected the simulation
components:
  schemas:
...
    MintSimulationInput:
      type: object
      properties:
        from:
          type: string
          pattern: "^0x[a-fA-F0-9]{40}$"
        data:
          type: string
          pattern: "^0x[a-fA-F0-9]+$"
          maxLength: 4096
      required:
        - from
        - data
    MintSimulationResult:
      type: object
      properties:
        result:
          type: string
          pattern: "^0x[a-fA-F0-9]*$"
      required:
        - result