# lib/api-zod/src/generated/types/mintSimulationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MintSimulationInput {
  /** @pattern ^0x[a-fA-F0-9]{40}$ */
  from: string;
  /**
     * @maxLength 4096
     * @pattern ^0x[a-fA-F0-9]+$
     */
  data: string;
}