# lib/api-zod/src/generated/types/mintSimulationResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MintSimulationResult {
  /** @pattern ^0x[a-fA-F0-9]*$ */
  result: string;
}